Update: Add handlers and adjust chatgpt service
In order to afford some more layers of abstraction, I added the handlers
folder to manage the business logic and, where necessary, access the
necessary services to handle requests.

The routes simply expose the handlers for consumption.

I also edited the chatgpt service to return response, not just print
it in console.